requirements_bot.txt
python-telegram-bot==20.2
httpx==0.23.3
